adr: mutualising filters
The goal of this ADR is to suggest the refactoring of RBH filters. The
idea is to mutualise the way filters are parsed through the command line
and generated to avoid code duplication in RBH tools.

Also, this ADR offers to avoid users to select which filters they need
to load as the information is directly stored in the mirror backend. In
our current implementation, it mutualized `rbh-find` and `rbh-lfind` in
the same tool without any user input.

Change-Id: Ib04f628e2250e6735c83382d0ba04544bb52cce3
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>