librobinhood: reorganize backends
Since the addition of extensions, a plugin can
either implement the backend or extension API.

This means that the backends directory doesn't
only contain backends therefore it is renamed
plugins.

To make it more explicit that the mfu, lustre
and retention extensions are posix specific,
they are moved inside the posix/{enrichers,iterators}
directories.

Change-Id: I0810bcf63c09fcef42c3077b3fdcefd5711661ec
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1211782
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>