find: add list of plugins and extensions to context
This patch adds a list of plugins and extensions to the find_context in
rbh-find. The goal will be to retrieve from the targeted mirror system
the list of backends used to create it, and call their predicate parsing
and filter building callbacks.

Currently, this list is filled with default values corresponding to the
tools usage. This will be changed later when the backends are stored in
the mirror system.

Change-Id: Ic932e2d3707c77714684a76fc9ccf5613cafabc9
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1211819
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>