find: add list of plugins to context
This patch adds a list of plugins to the find_context in rbh-find. The
goal will be to retrieve from the targeted mirror system the list of
backends used to create it, and call their predicate parsing and filter
building callbacks.

Currently, this list if filled with default values corresponding to the
tools usage and the plugin names given in the targeted URI. This will be
changed later when the backends are stored in the mirror system.

Change-Id: Ic932e2d3707c77714684a76fc9ccf5613cafabc9
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>