rbh-info: remove hardcoded path in test info.bash
The test was written so that the build directory of robinhood had to
be at <project_root>/builddir.

This is an invalid assumption so rely on the current working directory
from where the tests are run instead. This is also how we find
executables so this assumption is more correct.

Also, prevent the test from running in parallel with other tests
as it removes the mongo library from the build dir.

Change-Id: I41c2ed67b0746f9a3d4eccdcb9cc914d82174a6c
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1212115
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>