posix: fix invalid comparison in case of branch
This patch fixes an issue with an assert that would compare the root
given for synchronisation with the path given in case of branch.

When we do a rbh-sync with a branch, we will have the following:
```
rbh-sync rbh:lustre:a#b ...
```
Here, `path` is equal to `$PWD/a/b` while `root` is equal the realpath
of `$PWD/a`. Then, the assert would check `path` up until the length of
`root` is equal to `root`.

The issue comes if we do a sync with a `.` or `..` in it, for instance:
```
rbh-sync rbh:lustre:./a#b ...
```
Here, `path` is equal to `$PWD/./a/b` while `root` is equal the realpath
of `$PWD/./a`, which is `$PWD/a`. Thus, the assert would check that
`$PWD/a == $PWD/.` and error out.

This is simply fixed by getting the realpath of `path` before doing the
assert.

Change-Id: If2bbc72c0c6c6700a388c6764a9850b18c92dd63
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>