branch: fix an issue with Lustre branching
This patch fixes an issue where if we did a Lustre sync with a
branching, we wouldn't retrieve any Lustre information. This is due to
the Lustre branch backend creation function simply using the same as the
POSIX one. Thus, we would use the `iter_new` function of the POSIX backend,
and not the Lustre one, meaning the callback for retrieval of Lustre
info wouldn't be set, and would never be called.

Change-Id: I6b6c3bc4b0cfa8beaf6ca1fa667dc59731f717a7
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>