find: use the plugins delete callback for -delete
This patch changes the `-delete` handling of rbh-find to instead use the
callback from the plugins and extensions. Each plugin's callback is
called, but the first one to send back a success rc ends the loop.

The unlink is kept specifically for the MPI plugin, until the proper
callbacks are implemented.

Change-Id: I995137f2acd53d2d5383027f3c6bee6c09911975
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1212303
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>