utils: move time_from_timestamp to librobinhood
This patch moves the `time_from_timestamp` function from rbh-find utils
to librobinhood utils, as it will be used by the POSIX `print_entry`
callback later.

Also remove the utils files from rbh-find as, without
`time_from_timestamp` and `shell2pcre` (which was moved way before but
we forgot to remove the header prototype), the files are empty.

Change-Id: Ib08f1621399476b4c0d8aeb71a055f97a02f8b39
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>