rbh-report: fix inode based tests
Some tests of rbh-report look at the min/max of the
inode number between a few files. Filesystems do not
necessarily guarantee that inodes are created contiguously
or that newer files/dir will have larger inodes. For
instance, the inode number of a Lustre file is computed
based on its FID. So the inode number depends on the server
where it resides on Lustre.

Use find | sort to compute the min/max inode instead.

Change-Id: I2f5da619c17d6b424b5df39ef8ab0313bb890386
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1212432
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>