posix: retrieve the number of children in directories
This patch counts the number of children in directories. The count is
stored in the inode xattrs of a fsentry. When entering a directory in
post-order, a fsevent is generated to update the directory with the
right number of children.

The number of children in directories is stored in a stack. When
entering a new directory, the current counter is pushed onto the stack
to save its value, and we begin computing the number of children for
the current directory. When exiting this directory, we pop the stack
to retrieve the counter for its parent.

Change-Id: Ic9f51881d68302d753d97034209e2a46ae7ef66a
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>