mfu: retrieve the number of children in directories
This patch retrieves the number of children in directories for the mfu
iterator. The current parent’s path, ID, and number of children are
stored in static variables.

For each entry, we retrieve its parent. If it matches the current
parent’s path in the static variable, we increment the number of
children. Otherwise, if it differs, it means we are iterating through
a new directory. In this case, we generate a fsevent to update the
number of children for the previous directory and update the static
variables with the new parent’s path and children count.

With this method, we generate an additional fsevent per process if
all entries of a directory are listed together, which is typically the
case when iterating a file tree with mfu. However, this is not
necessarily true for an mfu file, as it can be sorted differently.
There is no optimal way to determine if we have seen all the children
of a directory without iterating through the entire mfu file.
Consequently, an fsevent is generated each time we encounter a new
directory during iteration.

Change-Id: Ia1148833c3aeb4bc450330642e491461700302d1
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>