tests: add db_cmd_sqlite
This command emulates the behavior of mongo queries but translate them
to sqlite queries. This script will make it possible to share the
code of tests between both backends without having to duplicate code.

This first patch makes the test test_posix_db_sync pass with both
the mongo and sqlite backend.

Change-Id: I231d02bb9d8116752ad544cf440a6b0159ba820d
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>