tests: replace mongo in URI for all tests
This patch replaces the name mongo with the variable
$db in all tests. The value of $db will be set to
mongo by default but can be overwritten by the environment
variable RBH_TEST_DB.

This is done so that all the tests can be run with SQLite
as a backend.

The only tests that can't be run with SQLite are the tests
doing concurrent write to the backend as SQLite only supports
one writer.

Change-Id: I2c5c8c3d165e5cb6ee3d065ac999fadc092e30e9
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>