find: move plugin imports to specific function
This patch moves the import of plugins to a specific function. It also
now uses the backend name requested in the URI to know which plugin to
import, now that the mpi-file callbacks are available. We now assume
that a Mongo URI will contain POSIX info and a MPI-file URI will contain
MPI info.

One thing to note is that even though the Mongo DB may contain information
from a MFU file, they use the same filters and actions, so filters that
apply to POSIX will also apply to MPI-file, and thus it's not an issue
if we don't use the MPI-file plugin for Mongo DBs.

Of course, this will all be simplified once source backends are
available in the Mongo DB and MPI file.

Change-Id: I3451a61fe19f1ab5df691813c5599cc8a4eed922
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>