lfind: remove obsolete functions from find and lfind
This patch removes the obsolete functions from rbh-find and
rbh-find-lustre. These are obsolete because they are now in the original
plugins, and only those are called.

These functions include part of the filters.c, actions.c and parser.c
files from rbh-find, and all the source files from rbh-find-lustre.

The rbh-find-lustre tool must be kept though, until we retrieve the list
of source backends from the DB, as we have no way of knowing the DB
contains Lustre information otherwise.

Change-Id: I5a32b51e0b9dcbb523b0005834a68b4f8490bef5
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1212483
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>