tests: fix archive_file test function
This function asks for the archival of a file and then waits for the
archived state to be set on the file. It was expecting the string
"archive_id:" on the output to consider the file as archived. This
behavior is not correct as the HSM attribute with the archive id is
written when the HSM archive request is received by the MDT not when
the file is actually archived. Therefore, we have a race condition
between the end of the archival and the call to "lfs hsm_state". If we
get the state after the end of the archival, the file is indeed
archived. But if we call it before, we can see archive_id in the
output but the file is still not archived.

This behavior results in the subsequent calls to lfs hsm_release to
fail since a non archived file can't be released.

To fix this, we need to wait for the archived state to be displayed on
the output. Since "lfs hsm_state" adds the filename to its output, we
need to remove the beginning of the output where the filename is because
some tests name their file "archived".

Change-Id: I0728ec18c94009f9410231d5e1527e9d6b64ed9c
Signed-off-by: Courrier Guillaume <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1212433
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>