posix: keep extensions instead of just enrich functions
This patch changes the posix backend structure to keep around the
extensions loaded instead of just the enrich functions. They will be
used later to record the plugin and extensions used as source backends
for a synchronisation, so that a call to rbh-info will show how a
database has been constructed.

Change-Id: Ic6687d9a1778372f0531a583c4569d9ab1cc025e
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1212999
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>