WIP rbh-policy: Provide initial code and example for policy engine
This patch introduces a Python file containing the initial
implementation of the policy engine command and another
Python file that serves as a sample configuration.
The goal is to highlight the engine's early functionality and
clarify the interaction between the command implementation and
a Python-based configuration in the design phase.

Change-Id: I1c4d69f21abdd889f9e995596eb8add63d72f803
Signed-off-by: Arnaud GAUTHERET <arnaud.gautheret@cea.fr>