librobinhood: move generic branch code
Mongo's implementation of the branch method is almost generic.
Only an explicit call to the implementation of rbh_backend_filter
method is done in this part of the code. This is done to avoid
infinit recursions since .filter would call rbh_backend_filter.

To prepare for the implementation of branch in the sqlite backend
this code is moved outside of the mongo backend. Since no backend
specific code is in it (apart from mongo_backend_filter), it can
be moved to librobinhood's sources.

Change-Id: Icb05474df579caca7b63a5f269df9d5542dabb2f
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>