s3: Add sync tests
This patch adds sync tests between s3 and mongo.
It begins by generating the credentials and starting the MinIO server in https.
After all tests are completed, the server is closed and deleted.
This patch also adds a variable for the path to the certificate
in the config file to connect via https

Change-Id: Ic4da377a08821600c5491e16b55e5f3037f4e26f
Signed-off-by: Alban RIO <Alban.RIO@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1213519
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Tested-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>