s3: Add sync tests
This patch adds sync tests between s3 and mongo.
It begins by generating the credentials and starting the MinIO server in https.
After all tests are completed, the server is closed and deleted.
This patch also add a variable for the path to the certificates
in the config file to connect via https

Change-Id: Ic4da377a08821600c5491e16b55e5f3037f4e26f
Signed-off-by: Alban RIO <Alban.RIO@cea.fr>