posix: extension loading doesn't setup enrichers
This patch removes the enricher setup from the extension loading part of
POSIX, and moves it the `backend_new` POSIX function. The goal is to
have a generic way of loading POSIX extensions without actually using
them, as other parts of the plugin may want to load them (namely, the
helper).

Change-Id: I80852cb329736acffafed3c278b941b4d0234943
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1213628
Reviewed-by: Arnaud Gautheret <arnaud.gautheret@cea.fr>
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>