find: add optional argument to `--help`
This patch adds an optional argument to rbh-find's helper that should
correspond to either a plugin or a backend defined in the configuration
file. If the helper is requested without an argument, then we print
find's basic helper. If a plugin or backend is requested, we load either
the plugin itself or the plugin that is extended by the backend, and
then call its helper function. Then we print the result of this call.

Change-Id: Iffb0bf97ddae7acdb2ee1680f421603abca6727e
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1213632
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>