fsevents: add callbacks to retrieve and set the source backends
This patch adds a callback to the enrich and sink operations, which will
retrieve the list of source backends used to enrich, and insert them in
the destination sink, respectively.

We also define the wrappers for these functions, and create the shells
for both the POSIX enricher and the backend sink. The Hestia enricher
and file sink callbacks are both set to NULL, but that may change in the
future.

The ultimate goal is to mimic the sync source backend synchronization,
by calling these two functions when doing a fsevents, so that a rbh-find
on the resulting backend can use the needed predicates and directives.

Change-Id: I8c53b67c937b9101fb4b0ff8a02037aa3b0d350e
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>