fsevents: add source sync between enricher and sink
This patch adds the function to synchronize the source backends from the
enricher to the sink. It is only called if the enricher is requested and
built properly.

Also add back the retention test to the list of tests that should
succeed.

Also remove the str2size_t function, since str2uint64_t already exists.

Change-Id: I1474385fdc54a0e7e088c44a43a8c8e28cf96dba
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>