retention: add helper for setting an entry's retention date
This patch adds a helper for setting an entry's retention date. The
goal is to provide an easier way for users to set a retention date,
without having to rely on knowledge about the extended attribute to set
or about timestamps.

Only the retention date is set, meaning the attribute that may later be
used by RBH to determine an entry's expiration date.

Change-Id: I6ab74a12d0aac03edd2ea0b8071748ca30806a52
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1213967
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-by: Arnaud Gautheret <arnaud.gautheret@cea.fr>