filters: move parse_expression function
This patch moves the parse_expression function to librobinhood. For
compatibility, the function is temporarily kept in rbh-find.

The function can also accept a callback as an argument to customize
the handling of actions and sort predicates. This is necessary because
rbh-report does not support actions, and rbh-report and rbh-find
handle sort predicates differently.

Change-Id: I5f8600f0d98573f7f8a6f1b3d55b1f547daad108
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>