report: add filters and remove short options
This patch adds filter support to rbh-report. Additionally, short
options have been removed to avoid ambiguity with predicates. For
example, -o for the "OR" predicate could be misinterpreted as the -o
option for --output.

Also, like with rbh-find, the --help, --config and --dry-run options
are now pre-uri arguments.

Finally, the helper has been modified to display the predicates
corresponding of the plugin used.

Change-Id: I438822f4312cb8c108e6ef82b4610f602750730e
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>