filters: add command_context structure
This patch adds a `command_context` structure to hold command related
options, including the config path, dry-run, helper and verbose. With
it, it adds to rbh-find the function to fill the structure depending on
the arguments given to the command, so that each field is filled, and
there is no need to pre-parse for multiple values.

Also add a `rbh_load_config` function to load a config from a given
config path, and remove the unnecessary argc/argv movement functions
from the config file.

Change-Id: Ie96225d59e48b1f42e4e0fa8d93940ac950a06aa
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>