info: now -s and -a options print a formatted size.
rbh-info -s and -a printing functions have been modified to
print a formatted size. Also, get_info tests have been
changed since the formated size are not matching the raw size
stored inside mongo. To do so, we use a format function to
format the raw size inside the test and compare it with the
result of rbh-info -s or -a option. This format function is
limited to KB

Change-Id: I5413d4d42db7331a6dd9077ffbd4f3914ba36b5d
Signed-off-by: Karl ENGELBRECHT <karl.engelbrecht@cea.fr>