mongo: add a refresh of the session
This patch creates an explicit session when performing a filter and
attaches it to the aggregate operation. Additionally, the session is
refreshed every 5 minutes to keep it alive.

This is useful when running a filter that lasts more than 30 minutes.
Otherwise, MongoDB will shut down the session and the filter will
fail.

Change-Id: Ic8cd69cae738f85d5f73bc0987187fa101430c71
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>