mpi-file: disable test on parent ID
There is a bug in the mpi-file backend on entries
that are at the root of the sync'ed directory.
dirname(path) returns '/' for both the (path = "/")
and direct children of the root (e.g. path = "/file").

The current implementation checks the result of dirname
to asses whether the entry is in the same directory as the
previous one. This is wrong for entries in the root.

This patch disables the related test to let the whole test
suite pass.

Change-Id: I6ec26828e6657574bd4c11159880cfdb9c258065
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>