tests: remove tests on layout generation
The layout generation is currently retrieved from the
"trusted.lov" xattr. In Lustre versions prior to 2.16,
the layout generation was always set to 0 by Lustre.
In newer versions, it is now set to (u16)-1 (e.g. 0xffff).

This makes all the tests that check the layout version fail.
The proper way to retrieve the version is through an custom
ioctl, not the xattr.

The layout generation tests are therefore disabled until we
properly fetch the layout.

Change-Id: I8036c4a0b50bae639c875251ae65b2e48d5ffda4
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>