sqlite: batch inserts in SQL transactions
Batching inserts provides a significant performance boost as SQLite does
not have to make sure that the data is safely stored on the disk after
each insert.

The synchronous mode is also disabled to let SQLite use asynchronous
I/O when writing on disk. This lets the application continue to work
instead of waiting for disk I/O.

Change-Id: I9693fa8e664e627254b092bad16d839e6fd73475
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1216666
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>