utils: add db_cmd_sqlite
This command emulates the behavior of mongo queries but translates them
to sqlite queries. This script will make it possible to share the code
of tests between both backends without having to duplicate test code.
This will make the maintenance of both backends easier in the long term.

With this patch, all the test can run successfully with the sqlite and
mongo backends. Except for the tests of rbh-report and rbh-info.

Change-Id: I926b2972c202265adc1ee1d844c4693e8d25148d
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1213458
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>