sqlite: batch inserts in SQL transactions
Batching inserts provides a significant performance boost as SQLite does
not have to make sure that the data is safely stored on the disk after
each insert.

Also use journal_mode = MEMORY to let SQLite store its journal in
memory. This also increases insert speed but can result in loss of
data if a crash happens (either a system crash or rbh-sync) in
the middle of a batch. Given the nature of robinhood, this is not
an issue.

The synchronous mode is also disabled to let SQLite use asynchronous
I/O when writing on disk. This lets the application continue to work
instead of waiting for disk I/O.

The last two parameters don't have an impact on performance as
significant as the transactions.

Change-Id: I9693fa8e664e627254b092bad16d839e6fd73475
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>