librobinhood: add basic allocation functions based on sstack
rbh_sstack_alloc allocates a given amount of memory on the sstack
that respects the memory alignement constraints of the system.
This will make it easier to implement generic allocation functions
on top of the sstack.

rbh_sstack_pop_all resets all the rbh_stack allocated in the sstack
but do not free the allocated buffers. This will avoid the reallocation
of memory inside iterators for instance. The first iteration will
allocate buffers that the subsequent iterations will reuse.

Change-Id: I11d3b14ded56be50c7c000f798cf1a5d5e78b394
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1216664
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>