ldiskfs: implement inode scan step
This first step scans all the inodes of the filesystem and stores them
in the dentry cache. The inode struct is duplicated as it contains the
metadata of the entry that will be used later for filtering. It also
reads the block numbers of directories and stores them in the directory
block list so that dentries can be scanned later.

Change-Id: I7b2076b3a02cf194c2fd6f5a35f8e58aa453175d
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>