sqlite: store build version when creating DB
To allow for future changes to the schema, the sqlite
backend now stores the build version in the database
and reads it when it is reopened after creation.

The version table is removed in favor of the info table.
Since both table only have a single entry, we can store
the version there.

Change-Id: Ib2bf8c724676fd197ce2aa10b1ed0c3c2573d9ae
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1216672
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>