mongo: Set INT32_MAX as socket timeout
This patch sets the libmongoc socket timeout to the maximum possible
value, INT32_MAX, which corresponds to a duration of over 24 days.

By default, the timeout is set to 5 minutes, which can cause timeouts
when calling cursor_get_next(). With rbh-report or rbh-find using a
complex filter on a collection of more than 100M documents, it is
highly probable that a getMore operation on the MongoDB side will take
longer than 5 minutes to complete.

Change-Id: I044dba5e857fcadb036fea5a24299fa662300162
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>