find: initialize the projection
This patch initializes the projection with the information needed by
rbh-find. For example, we set the projection to retrieve only the path
if we are using the print action.

For now, the printf and fprintf actions are not yet handled. The
projection will be initiliazed from the format string in later patches.

Change-Id: Ic66fb344c896fe3d5b03b2296426d7024d611e67
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>