find: initialize the projection for printf/fprintf action
This patch use the fill_projection callback of the plugin/extension to
initialize the projection to use when using printf/fprintf action.

Also, removes the projection to false when the projection mask is zero
because Mongo doesn't support having a projection with field at true
("form" : "fsentry") and false at the same time.

Change-Id: I10cb8d6ca6301018d2d52f9cc121f788109a41fe
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>