lustre: use an ioctl instead of xattrs for the layout gen
This patch adds an ioctl to retrieve the layout generation instead of
using the trusted.lov xattrs. In earlier version of Lustre, the layout
generation is not properly stored in trusted.lov for a file with a
simple layout. To retrieve the right value, we need to use an ioctl.

Change-Id: I45fa098e5168ed0a49fab3f376d73eca6a79e641
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>