rbh-fsevents: fix parsing of MDT name
This patch fixes the parsing of the MDT name, which was failing when
the Lustre filesystem name contained a number. Previously, the code
searched for the first occurrence of a digit and attempted to convert
the rest of the string to an integer. For example, with a filesystem
named blob1 and an MDT named blob1-MDT0000, it would incorrectly try
to convert 1-MDT0000 to an integer.

The parsing logic has been reversed to search for the first occurrence
of a character from the end.

Change-Id: I113b5ec64073b0bee814ea7595fcbeaafefc60e7
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-on: https://gerrithub.io/c/robinhood-suite/robinhood4/+/1218613
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>