lustre: don't retrieve layout information for block/char devices, FIFOs, or sockets
This patch stops retrieving layout information for block and character
devices, FIFOs, and sockets, as there is no information to retrieve
(e.g., lfs getstripe). This change prepares for the next patch that
will modify the method of retrieving layout information via ioctl.

Also, this patch adds the FID and MDT index to the xattrs of
block/char devices, FIFOs, and sockets, as is already done for
symlinks.

Change-Id: I47a7a45d4a37b8b346637b61c7827eca1a19a1ee
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>