rbh-fsevents: open only once per entry
This patch externalizes the file descriptor associated with an entry,
from the enrich function to the enricher structure. This allows the
same file descriptor to be reused for all fsevents related to that
entry. The file descriptor is only closed when the fsevent to enrich
pertains to a different entry.

Change-Id: I5302ad368155d4bbaa0da0cedea7f5e3da42b786
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>