rbh-fsevents: ensure the last possible fsevents are deduplicated
This patch ensures that all consecutive fsevents for a single entry
are deduplicated. Previously, when there was one slot left in the
batch and a new entry’s fsevents were found, the batch was marked as
full and flushed, even though additional fsevents for the same entry
followed immediately. As a result, the fsevents for the last entry in
the batch were always split across two batches.

Change-Id: Iac43af48d03000d0272ebd1647f610d1a488e642
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>