rbh-fsevents: ensure the last possible fsevents are deduplicated
This patch ensures that we do deduplication until we encounter a new
entry. Previously, when there was one slot left in the batch and a new
entry’s fsevents were found, the batch was marked as full and flushed,
even though additional fsevents for entries inside the pool followed
immediately.

Change-Id: Iac43af48d03000d0272ebd1647f610d1a488e642
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-on: https://gerrithub.io/c/robinhood-suite/robinhood4/+/1221159
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>