rbh-policy: Added the function that will be used to create the filters
Added the function that will be used in python to create filters.
This function takes the uri as the arguments of the filter to be
created and uses the codes from the previous patch to create a
filter.
If the uri does not have this filter, it returns NULL, otherwise
it returns the filter.

Change-Id: Ibcaabdb887e00078ec142975107a749378cf78f3
Signed-off-by: Arnaud GAUTHERET <arnaud.gautheret@cea.fr>