librobinhood: add a callback to free elements in array and list iterators
This patch adds a callback to the array and list iterators to free
elements when destroying the iterator. The callback function can be
provided at the creation of the iterator, or omitted by specifying
NULL. This will make it easier to free the memory used by array and
list iterators inside rbh-fsevents.

Change-Id: I3ca8c7d62fd6a986cd9d4ef02683be6ae0a68083
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-on: https://gerrithub.io/c/robinhood-suite/robinhood4/+/1221160
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>