rbh-fsevents: add hashmap to track entries in process
This patch introduces a hashmap to track entries that are currently
part of a batch being enriched. This mechanism allows us to determine
whether an entry can be sent for enrichment or not.

 - If an entry is already present in the hashmap, it is enqueued in a
   waiting list.
 - If an entry is not in the hashmap, it is added and processed.
 - Once a batch has been enriched and updated in MongoDB, all its entries
   are removed from the hashmap.

Before flushing entries from the dedup, the waiting list is iterated
over to check if any entries can now be sent for enrichment.

This ensures that no two parallel batches process the same entry
simultaneously.

Change-Id: I59c1166afbdcd5d9988d49c7bf121e5007663c6a
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>