rbh-fsevents: duplicate fsevent arrays and lists
This patch duplicates the fsevent arrays and lists retrieved from
dedup to prepare for parallelism in rbh-fsevents. This is necessary
because each batch will be stored in a queue, waiting to be enriched
by other threads, while dedup frees the contents of the array or list
immediately after starting a new batch. Without this, the contents of
the batches in the queue may be freed before being enriched.

Change-Id: Ia84a7f105bdf6108eb2ab19f2202fad7f703ae93
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>