rbh-fsevents: parallelize enrichment and updates in Mongo
This patch parallelizes enrichment and updates in Mongo using a
producer/consumer model. The producer generates batches of fsevents to
be enriched and updated in Mongo, while a pool of consumers processes
each batch in parallel.

How it works:
The producer reads the changelog and generates a batch of fsevents.
This batch is enqueued, and the producer continues generating the next
batch until there are no more changelogs to read. Consumers dequeue
batches, enrich them, and update them in Mongo until the queue is
empty and the producer has no more work.

Also, the 'nb-workers' option has been added to specify the number of
consumers to use. By default, there is one producer and one consumer.

Finally, there can be race conditions between batches. For example, if
batch1 contains a LINK fsevent and batch2 contains an UNLINK fsevent,
it is possible for batch2 to be updated in Mongo before batch1,
leading to inconsistencies in Mongo. This issue will be addressed in
upcoming patches.

Change-Id: I64d82a2bc222e3d70017a9017db72dec1b156cc8
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>