rbh-fsevents: make no-dedup mode compatible with multiple workers
This patch updates the no-dedup mode to work correctly with multiple
workers. Before sending an fsevent to enrichment, we now check whether
its entry is already being processed:
 - If the entry is in progress, the fsevent is queued for later
   enrichment.
 - If not, the fsevent is sent for enrichment and marked as in
   progress.

Finally, before processing a new fsevent, we check if any queued
fsevents are now eligible to be sent to the enrichment step.

Change-Id: I6d73ab9912cd93842f9099f6287c27b3a7c30ec5
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>