mfu: fix bug with last parent_id being NULL
This patch fixes a bug where a NULL pointer could be used,
causing a segmentation fault. A check has been added to avoid using
the pointer when it is NULL.

The last parent_id may be NULL if the directory is deleted between the
mfu walk and the Robinhood enrichment.

Additionally, this patch adds a test to verify that when the last
parent_id is NULL, the error is handled correctly.

Change-Id: I21d48d2bf99e16f6a94dfba9ca11d01634fc5334
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>