rbh-fsevents: distribute IDs across threads based on FID hash
This patch distributes the IDs of a dedup batch across all threads
based on the FID’s hash. This ensures that a given ID is always sent
to the same thread, resolving potential races between threads.

For each batch, one iterator per thread is now created. When flushing
the batch, the dedup returns an iterator that contains all per-thread
iterators associated with their thread index. The producer thread
simply iterates over this iterator and dispatches the work to the
corresponding index.

Change-Id: I0dd8f245f054d6c9807750a3069103564049e3f4
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>