undelete: retrieve mountpoint from filesystem
This patch changes the default mountpoint retrieval mechanism to instead
get it from the current filesystem. If it cannot be retrieved, use the
one given in the database.

To retrieve the mountpoint from the filesystem, we do the following:
```
FID = path2fid($PWD)
fsentry = filter(database, FID)
mount_point = $PWD - fsentry.path
```

Change-Id: I2e9d25da6c3fb491faec4966f4a63df60056ac4e
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1224046
Reviewed-by: Patrice LUCAS <patrice.lucas@cea.fr>
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>