fsevents: add acknowledgement of changelog during the commands
This patch adds the acknowledgement of the changelogs during
rbh-fsevents. The changelog are now ack batch by batch.
After completing a batch, the source will now keep in memory what is
the last changelog index of this batch. When all consumers have
finished to handle this batch, the source will now be able to
acknowledge this batch.

Two callbacks has been added to the source: one to save the last batch
return (save_batch) and one to ack a batch (ack_batch).

Change-Id: I2baaffb29eddc5bc00cfbd9a5244dd88ca06395c
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>