retention: add delay to rbh-update-retention
This patch adds a delay option to rbh-update-retention. Its goal it to
provide a time window during which directories that should truly expire
will be notified to the user. This allows them to update the directory
or its content if needed, or at least know beforehand what will be
removed.

Also modify the code around to be clearer for each case/step of
expiration.

Change-Id: I6205c11b408722ddbf41004f4ff9aba0766e9c5d
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1224461
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-by: Arnaud Gautheret <arnaud.gautheret@cea.fr>