sqlite: make fsevents and sync tests pass
Some more tests need to be skipped:
- concurrent inserts with MPI are not supported;
- the $exists filter is not yet implemented;
- skip tests with MPI. Since sqlite does not support
  concurrent inserts, we either have to run them with
  mongo or skip them entirely. It is preferable to skip
  them as it would be confusing to see a test run with
  mongo when specifying RBH_TEST_DB=sqlite.

Modify test_lustre_db_sync to pass an actual
list for the stripe count check. This is easier
to implement in sqlite and also works with mongo.

db_cmd_sqlite is updated:
- support both BinData and Binary.createFromBase64;
- support ns.parent filter;
- refactor type resolution and filter initialisation;
- improve the replacement of _id to only match _id and not mirror_id
  for instance;
- add support for json like arguments. For instance, JSON arrays for the
  stripe count;
- properly handle binary data by printing the decoded base 64 format;
- implement remove_entries.

Update the count and find query:

When a test uses filters that match two hardlinks of the same inode, we
need to make sure that we don't count them twice or that exclusive
filters still match the inode. For instance, an inode with hardlinks 'a'
and 'b' can be matched with the filter '{ns.name: "a"} {ns.name: "b"}'.
To achieve this, we first match all the rows from "entries left join ns"
and then group them by id to deduplicate the hardlinks. Finally, we
count the entries from that deduplicated table to get the final result.
This query is also extented to be able to return actual data instead
of just the count. This is used by the get method.

Change-Id: I91546c114ee8bd67973815a8bed0be223921284a