db_cmd_sqlite: add get_info and drop_info
To make rbh-undelete tests pass with sqlite, we need
to remove the explicit use of the mongo command in the
tests to retrieve the mountpoint. The command get_info
has been added in both db_cmd_sqlite and db_cmd_mongo
to achieve this. Also add drop_info used in rbh-undelete's
tests.

Change-Id: Ideb7e3bc3279c3b653ab9dd34152c5fcaa16894a
Signed-off-by: Guillaume Courrier <guillaume.courrier@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1225294
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Tested-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>