utils: remove `which` error message for mongo
This patch removes the error message shown when using the `which`
command to get the mongo command. Since the mongo command is either
`mongo` or `mongosh`, depending on the version installed, if the
installed one was `mongo`, the `which` error for `mongosh` would show up
in every test. But since it actually does not correspond to an error,
better just remove it entirely.

Change-Id: I250c75b8a373b5133c83a6e2fa278308e5a20cf9
Signed-off-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1225873
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>