gc: add option `--sync-gc`
This option takes an epoch as an argument. This epoch is used during the
backend retrieval to filter only entries which have a `ns-xattrs.sync_time`
lesser than the epoch. The default behavior of `rbh-gc` remains unchanged.

The `ns-xattrs.sync_time` is now updated with each synchronisation from
a POSIX backend.

A test is added to check the behavior of the option.

Change-Id: Icb68e3d7ca273c64d0a61ecef798c9c81bdf6188
Signed-off-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>
Reviewed-on: https://review.gerrithub.io/c/robinhood-suite/robinhood4/+/1229540
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>