rbh-update-path: add overload of rbh_iter_list
This patch adds an overload of rbh_iter_list to be used with a
structure containing a pointer instead of a static variable.
It will be used to store the fsevents and fsentries of children.

Without this overload, if a structure node contains a pointer,
rbh_iter_list returns a pointer to that pointer, which cannot be used
with rbh_backend_update.

Change-Id: I0e794dc6df7c274a35a5bbf573fdf7af0a944bea
Signed-off-by: Gauthier EVRAERD <Gauthier.EVRAERD@cea.fr>
Reviewed-on: https://gerrithub.io/c/robinhood-suite/robinhood4/+/1227841
Reviewed-by: VALERI Yoann <yoann.valeri@cea.fr>
Reviewed-by: Sebastien Gougeaud <sebastien.gougeaud@cea.fr>